Commerce module v1.0.0.9

+++ commerce/interaction-api/src/main/java/ru/yandex/practicum/commerce/interactionapi/dto/BookedProductsDto.java
@@ -5,15 +5,17 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder(toBuilder = true)
 @AllArgsConstructor
 @NoArgsConstructor
 public class BookedProductsDto {
 
-    private Double deliveryWeight;
+    private BigDecimal deliveryWeight;
 
-    private Double deliveryVolume;
+    private BigDecimal deliveryVolume;
 
     private boolean fragile;
 }
